Get willow and pytorch to do the same reduction type for NLL loss
Summary:
The default loss for pytorch NLL is the mean over batches.
Willow uses a sum of batches. We tell pytorch to take the mean over
batches too. As this is the first case with serious debugging, I
took a few notes and added those to the repo. Also, fixed the incorrect
file used for the numerics report.

Test Plan: added a new test model5.py

Reviewers: davidn, evan, georgep

Reviewed By: georgep

Differential Revision: https://phabricator.sourcevertex.net/D7366