Separate the scheduler to allow for a selection of different ones
Summary:
Separate out the function which creates the ordered list of
operations for creating the graph (the schedule).

Put it into a 'TopologicalScheduler' class.

Add a factory for getting different schedulers.

Test Plan:
All existing tests ok.

follow up task to allow user to select the scheduler should add a check that the schedule selection works.

Reviewers: jakeh, georgep, simonl, jamesn

Reviewed By: jamesn

Differential Revision: https://phabricator.sourcevertex.net/D7427