Add a getter for supported ops
Summary:
Adding a getter for the list of all ops.

have refactored the OpTypes class into a separate file and attached
the domain to the op so that they make a pair as the spec says.

have added a boolean for 'is public' to the list.

Test Plan: New test added for getting the list.

Reviewers: jamesn, george

Reviewed By: george

Differential Revision: https://phabricator.sourcevertex.net/D7706