Removed batchSize input of DataFlow constructor, changed anchor array shape to align with what CE expect
Summary:
Motivation for diff:
1. The batchSize member of DataFlow class was unused in poponnx.

2. After consulting with DavidN and CE, we are changing the spec for output data (anchors) for the API.

To clarify what we expect for the input array to a step:

Single-sample inputs do not have a 'batch' dimension
- Sample shape e.g. : [3, 32, 32]
- Batch shape, batch-size = 1 : [3, 32, 32]
- Batch shape, batch-size = 4 : [4, 3, 32, 32] (only if parallelizing across outer dimension! As @davidn ponits out, this is not the case for LSTM models)

We add an outer dimension for step-size when supplying an input array to the stepio (unless step-size == 1, in which case step shape = batch shape)
- Step shape, step-size = 1, batch-size = 1 : [3, 32, 32]
- Step shape, step-size = 2, batch-size = 1 : [2, 3, 32, 32]
- Step shape, step-size = 2, batch-size = 4 : e.g. [2, 4, 3, 32, 32]

When copying anchors to the host after a step, we previously returned an array whose outer dimension was equal to (step-size * outer dim of batch shape).
Now we return batches over the shape size. The shape of the anchor array has an additional dependency on anchor return type, and anchor return frequency (if applicable).
- Anchor tensor shape e.g. : [2, 32, 36]
- Anchor array shape, step-size = 1, art = ALL : [2, 32, 36]
- Anchor array shape, step-size = 4, art = ALL : [4, 2, 32, 36]
- Anchor array shape, step-size = 4, art = FINAL : [2, 32, 36]
- Anchor array shape, step-size = 4, art = EVERYN, N = 4 : [2, 32, 36]
- Anchor array shape, step-size = 4, art = EVERYN, N = 2 : [2, 2, 32, 36]

See anchorreturntype_test.py to check that poponnx is happy with this spec.

Diff summary:
Fixes T5750
Changed spec for input and output tensor shape for poponnx. Changed host stream shape, anchor array shape, and fixed examples

Test Plan: Tests pass. New tests added to anchorreturntype_test

Reviewers: davidn, georgep, jakeh, jamesn

Reviewed By: jakeh

Subscribers: davel, oskarb, markp, davidn

Maniphest Tasks: T5750

Differential Revision: https://phabricator.sourcevertex.net/D8028