Don't mutate graph names unnecessarily
Summary:
When loading a graph, the names of the input/output/initializers
were being unnecessarily corrupted.  The translation table was not
enough to indicate which of the original names was the translated
one in the case where two graphs were merged (by a future feature)

Removed the uniqifier.

Also added a min and max supported opset version.  Probably a
placeholder for some future opset registration mechanism.

Test Plan: All tests ok

Reviewers: simonl

Reviewed By: simonl

Differential Revision: https://phabricator.sourcevertex.net/D8567