Copy attributes and debug context when replacing ops in patterns
Summary:
Methods added to Pattern class to copy name, virtual graph and recomputation attributes are copies from op to be destroyed to ops to be created.

Built into this change is that a lot of the boiler plate code is moved from the pattern implementation into the `makeReplacementOpInIr`. All patterns refactored to use the new method.

Inside this method, we now use `OpManager` to create new ops. The downside of this is that it uses the `Op` constructor.
So for cases when we want to set class members when creating a new op (e.g. `Scale` or `ReduceSum`) we do a `dynamic_cast` to the Op type, then use setter functions.

The alternative way of setting class members that I considered was by using onnx attributes (i.e. using the same way the members are set when creating the Op from an onnx model), as follows.

   auto attr = generateAttributes({"keepdims", "axes"});
   attr.set(keepdims, "keepdims");
   attr.set(axes, "axes");
   auto reducesum = makeReplacementOpInIr(Onnx::Operators::ReduceSum, op, attr);

However, I used the first approach here. Seems neater not to use onnx protobuf attributes once inside the Ir.

Fixes T6061

Test Plan: Added cpp pattern test. All tests pass.

Reviewers: simonl, jamesn, jakeh

Reviewed By: jamesn

Maniphest Tasks: T6061

Differential Revision: https://phabricator.sourcevertex.net/D8640