Added a custom allocator for `GatherOpx`.
Summary:
Based on the TF dynamic slice add tensor D6859.

>>! In T4746#84194, @davidn wrote:
> See poplibs dynamic slice test.
>
> actual allocation should be:
>
> [BxN/G, S, G]
>
> where B is batch, N is channels, S is the sequence length, and G is chosen so that BxN is not larger than the number of tiles.
>
> The tensor should then be linearly mapped by slicing along the sequence dimension and then doing a linear map on each slice.
>
> finally the allocated variable should then be dim shuffled into whatever form TF wants.

I've taken this approach and also added a balance factor `H`.
So I will create a tensor of shape `[DxH/G, S/H, G]`, where `D` is the product of all dimensions outside of the gather axis dimension (`BxN` in the previous description).
This will put `SxG/H` elements on each tile.

This balance factor will allow us to handle tensors with a very large gather axis.
For example 73903x512 gathered on axis 0:
```
D = 512
S = 73903
```
This would mean `G = 1` and `H = 2`, so the allocated tensor would be `[512x2/1, 73903/2, 1] = [1024, 36952, 1]`, obviously some padding will need to be used in this case.
This will place 36952 elements on each tile.

Resolves T6585
Depends on T6571

I suspect there's a desire to move this into poplibs in the future.

Test Plan:
I was able to run a gather test case of size [73903, 512].
I've not added this to the set of tests because it takes a very long time to run.

Existing tests pass and builds with -Werror

Reviewers: jamesn, davidn

Reviewed By: davidn

Maniphest Tasks: T6585

Differential Revision: https://phabricator.sourcevertex.net/D8952