Inplace & view-changing
Summary:
The new algorithm keeps track of which regions of tensors are aliased with
regions of other tensors, and which regions of an op's input are modified.
This information is propogated through the graph to detect more precisely
if and when ops can be in-place or view-changing (as opposed to requiring new
device memory). New classes to achieve this are,
- Region : Slices/sub-rectangles of a tensor
- Link, Chain, Chains : used to describe how Regions get mapped by Ops.

Test Plan: 2 new tests added, many more coming in T6707 (follow up task)

Reviewers: davidn, evan, george, georgep, jakeh, simonl

Reviewed By: jakeh

Differential Revision: https://phabricator.sourcevertex.net/D9063