Changes for builder opset support
Summary:
This diff has the complete set of changes

For C++

The builder class now exposes classes per domain/opset with the matching operations
The onnx operations are defined in the opset that they are introduced and then 'pull' into the later opset via the 'using' statements

Example
```
auto builder = Builder::create();
auto aiOnnx = builder->aiOnnxOpset9();
auto o = aiOnnx.add({i1, i2});
```
```
auto builder = Builder::create();
auto aiOnnx = builder->aiOnnxOpset6();
auto o = aiOnnx.add({i1, i2}, 0, 1);
```

For Python

When the builder class is created it takes a dictionary of domains to versions. This information is then used when the user asks for the interface for a domain and the build class returns the right version
The mess bit is the pybind code which needs  the complete opset definition - any ideas on how to do that better appreciated.

Example - The 'Builder()' will return opset 9 by default
```
builder = poponnx.Builder()
o = builder.aiOnnx.add((i1, i2))
```
```
builder = poponnx.Builder({"ai.onnx":6})
aiOnnx = builder.aiOnnx
o = aiOnnx.add((i1, i2), 0, 1)
```

Additionally the majority of the builder interface and pybind code is autogenerated using the gen_operators.py script using the ONNX operator definition from the onnx repo. This is the same definition which is used to build the ONNX operator documentation,

The script generate files with the suffix .gen. These are included into the files that are compiled.

There has been change to the order of parameters to the builder functions to match the order in the the specification. As a general rule if the parametere value is the onnx default it will not be included in the attribute list of the attribute. For lists, they will not be included if empty. In a few cases there are optional attributes that do not have default values, they are now represented as boost optionals.

Test Plan: ./tesh.sh all - passes

Reviewers: davidn, jakeh, evan, george, jamesn

Reviewed By: davidn, jakeh, evan, george, jamesn

Subscribers: georgep

Maniphest Tasks: T6274

Differential Revision: https://phabricator.sourcevertex.net/D9066