Move getGradOp error to IR
Summary:
In autoVirtualGraph a gradOp is created to estimate the cost
of a backwards op. If this op does not have a gradOp then this error is thrown.
Ops may not have a gradOp but still be valid in the fwd pass if they do not contribute
to the loss. This a temp fix until the cost function can check if an op
contributes to the losses.

Test Plan: all tests pass.

Reviewers: jamesn, davidn

Reviewed By: jamesn

Differential Revision: https://phabricator.sourcevertex.net/D9951