Efficiency improvement for Inplacing Pattern
Summary:
Inserting empty Chains causes a quadratic growth of empty aliases in the
Slice Op. This diff postpones the creation of new chains until it is known
that they are not empty. We see a 3X speed-up for the benchmark test with
N=5, J=5. With N=5 J=5 after this diff, 35% of the time is spent checking
to see if the graph is topologically sortable after inplacing.

Test Plan: All tests pass.

Reviewers: george

Reviewed By: george

Differential Revision: https://phabricator.sourcevertex.net/D10702