Add constraints on ConvFlipWeights op, remove 'priority' setting in constructor
Summary:
Fixes T8423

Add constraint in Ir to ConvFlipWeights ops, so that they are run as late as possible. This means we can remove the `priority` setting in the Op's constructor. This had the same effect, but conflicted with the `priority` setting of the recompute ops, causing a badly optimized schedule for min liveness.

Test Plan: All tests pass. Preliminary frameworks_comparison test suggests that memory is back in-line with Enigma NormOnly recomputation.

Reviewers: jamesn, george, simonl

Reviewed By: jamesn, simonl

Maniphest Tasks: T8423

Differential Revision: https://phabricator.sourcevertex.net/D11067