Add ReductionType to losses
Summary:
See comment in loss.hpp:
```
When weight updates of a batch are computed in one go, we
are reducing over the gradients of the whole minibatch.
What type of reduction should this be?
SUM : By scaling the loss gradient (and loss) by identity,
      this is a sum reduction
MEAN : By dividing the loss gradient (and loss) by total
       number of samples this is an average (mean) reduction
```

The scaling by 'total number of samples' in the `MEAN` case includes samples per batch in a single graph AND replicated copies.

This is a backwards compatible change at the python API level. The `reduction` parameter has default `SUM`, but can be overridden.

Fixes T8834
Addresses T8836 (reduction is SUM by default, so as not to require an API change for this release)

Test Plan: All tests pass. Modified graph_replication_test.py and loss_test.py to use new mean reduction"

Reviewers: simonl, george, jamesn, matthewha

Reviewed By: simonl

Subscribers: davidn, dominicm

Maniphest Tasks: T8836, T8834

Differential Revision: https://phabricator.sourcevertex.net/D11668