Adding python tests via glob
Summary:
Python tests were not being added to the CMakeLists.txt, and so were not being run. This diff adds the cmake function `add_all_py_unit_tests` which should add a test for each python file in a directory.

A list of files to exclude may be passed as an argument for files such as `test_util.py` or tests failing for a specific reason.

Using the `CONFIGURE_DEPENDS` argument to `file(GLOB` means new tests should be detected without having to explicitly rerun cmake or rebuild the project.

Test Plan: All tests still present and passing

Reviewers: jamesn, davidn, evan, simonl, matthewha

Reviewed By: evan

Differential Revision: https://phabricator.sourcevertex.net/D11882