Added a test to check for memory regressions
Summary:
Test using code modified from the frameworks comparison resnet18_bs1_bn example,
trimmed down to run a bit faster. The total and max memory are compared against a reference,
and if it is 5% over or under, the test will fail.

The reason for failing the test if memory is sufficiently under the limit, is to notify us as to when the limit needs to be updated.

Reviewers: davidn, simonl, jamesn, evan, matthewha

Reviewed By: simonl

Maniphest Tasks: T8693

Differential Revision: https://phabricator.sourcevertex.net/D12382