Distinguish between Tensor creation and Tensor population.
Summary:
We have been using taskWhichCreates for 2 purposes. (1) The task which adds
the Tensor to a Graph and (2) the task which adds code to a program to set the
Tensor's value. Often (1) and (2) are the same task, for example poplibs API
calls both create a Tensor and generate code. But not always; a Stream Tensor
is only safe to use once populated, having been created is not enough.

Test Plan: A new test added. This test fails without the changes in this diff

Reviewers: simonl, george, evan

Reviewed By: evan

Subscribers: kateh

Differential Revision: https://phabricator.sourcevertex.net/D12447