Add recompute for pipelining at the Ir level
Summary:
With recompute enabled, a minimal sub-set of the stashing candidates is selected.
Candidates which are eliminated from the stashing proposal set have their
producers annotated as RECOMPUTE. The error thrown when both pipelining and
recomputation are enabled is pushed back to the Devicex level, so that Ir tests
can be performed.

Test Plan: 3 new Ir tests of increasing complexity are added.

Reviewers: evan, matthewha, simonl

Reviewed By: evan

Differential Revision: https://phabricator.sourcevertex.net/D13471