Convert MatMulXXXGrad to MatMul
Summary:
Added pattern to replace MatMulLhsGradOp and MatMulRhsGradOp with a Transpose & MatMulOp.

This is to facilitate the outling of matmuls and enable the option to group matmuls.

Had to change the code for tensor creation to not throw an exception when two op's claim they can create a tensor, not it just picks the first.....

Test Plan: ./test.sh all - passes

Reviewers: jamesn, evan, george, jakeh, matthewha

Reviewed By: george

Maniphest Tasks: T9766

Differential Revision: https://phabricator.sourcevertex.net/D13320