Lump all Restores together in a fragment, run between fdw and bwd
Summary:
Attempts to get Restores to be the first Ops in the bwd fragment
all result in unreasonably complex code. With the approach in this diff,
topological constraints between Restores and other Ops are not needed
(and if present, ignored) other than putting Restore post-loss.

I'm still not completely satisfied with this way of letting the backend ignore
topological constraints (used elsewhere, for example IPUCopyOps in pipelining u
se this trick). A clean and simple Ir approach should still be sought.

Test Plan: All tests pass

Reviewers: simonl, evan, matthewha

Reviewed By: simonl

Differential Revision: https://phabricator.sourcevertex.net/D13633