Added virtualGraphMode option
Summary:
Added the option virtualGraphMode.

I've left the options enableVirtualGraphs and autoVirtualGraph in so as not to break any of the public examples or frameworks comparisons,
but it now prints an 'option deprecated' warning if either are set.

If the user sets virtualGraphMode to something other than Off, enableVirtualGraph and autoVirtualGraph are ignored.

I haven't changed any of the tests to use this option yet. I would prefer to do that in a separate diff.

Reviewers: jamesn, evan, davidn

Reviewed By: evan

Maniphest Tasks: T10462

Differential Revision: https://phabricator.sourcevertex.net/D13814