Fix possible bug in constexpr
Summary:
It was assumed before that if the graph was not training, then const folding could be done on variable tensors as well as const tensors.
The issue is that the const folding is done once, when the session is created and if the user calls `resetHostWeights`, the const folding will not be performed on these new weights.

Test Plan: all tests still passing

Reviewers: oskarb

Reviewed By: oskarb

Differential Revision: https://phabricator.sourcevertex.net/D13938