Add custom codelet user option
Summary:
Slight modification of D12923.

I messed around with a few options for this, such as adding codelets via method in the session like `session.addcodelets` to enable it to be more dynamic, but it didn't really fit.

This at least allows the user to add multiple codelets in one go, and allow poplar to infer the file type. Also allows for compile flags. Only poplar functionality this doesn't allow for is the `ostream` output.

Test Plan: Tested with a codelet example which is logged as loaded in poplar logs, how would I test this without the logs?

Reviewers: simonl, jamesn, george, evan

Reviewed By: simonl

Maniphest Tasks: T9004

Differential Revision: https://phabricator.sourcevertex.net/D14151