Tensor specific optimizer-values and constness
Summary:
The SGD Optimizer now has

1) a loss scaling factor
2) global learning rate
3) global weight decay factor
4) optional learning rates and weight decays for all Variables

Any combination of the both can be be const (compiled into engine) or variable
(can be updated on device at any time).

One subtlety is that the user might specify a constant learning rate but a
non-constant loss scaling factor, in which case the scaled learning rate is not
constant. Similary, a non-constant learning rate implies a non-constant weight
decay scale factor, even if the weight decay is specified by the user as constant.

Accompanying changes,

- Removed SGDBuilder
- Removed BaseSGD
- ConstSGD is ready to be removed but kept for python API back compatiblity
- Optimizer Tensors (other than loss scaling Tensors) are created in Ir jit.
- Change LearnRate to LearningRate wherever found
- Change weight decay scale factor from (1 - wd*(lr/ls)) to (1 - wd*lr)

Follow-up task: expose API to python

Test Plan: - Tests of different combinations of const and non-const, global and specific

Reviewers: george, evan, simonl, matthewha

Reviewed By: simonl

Differential Revision: https://phabricator.sourcevertex.net/D14527