Added methods to retrieve the pipeline stage from the builder.
Summary:
This change enables cleaner example code for T11506, as we won't have to retain a separate member variable to track the pipeline-stage scope (as per D15393).

Added a general attribute getter to the builder in the C++ API, and opened a specific getter for the pipeline stage into the Python interface.

If the pipeline stage is not set, it will throw an exception - this seemed preferable to returning 0 as it's explicit that the stage hasn't been set.

Added a method to check if the pipeline stage has been set before retrieving it, to avoid unnecessarily throwing errors.

Test Plan: All tests still pass.

Reviewers: george, evan, jamesn, simonl

Reviewed By: simonl

Tags: #bert

Differential Revision: https://phabricator.sourcevertex.net/D15430