Chain copies to prevent duplicates being created by the contiguate copies transform.
Summary:
If a tensor is consumed on a different ipu, a copy will be added from the ipu the tensor is produced on, to the ipu the tensor is consumed on.
A copy is inserted for each ipu the copy is consumed on.
When the contiguate copies transform runs, for each copy, a chain of copies is created going to every pipeline stage between the source and destination.
This can result in duplicate copies. This:

```
stage0    |     stageN    |   stageM
  X -> [copy] -> X_c0
  X -> [copy] ---------------> X_c1
```

becomes:

```
stage0    |     stage1      |     stage2 | ... | stageN      |     stageN+1 | ... | stageM
  X -> [copy] -> X_n0 -> [copy] -> X_n1 -> ... -> X_c0
  X -> [copy] -> X_m0 -> [copy] -> X_m1 -> ... -> X_mn -> [copy] -> X_mn+1 -> ... -> X_c1
```

Before the contiguate copies transform, we chain the copies:

```
stage0    |     stageN      |     stageM
  X -> [copy] -> X_c0 -> [copy] -> X_c1
```

resulting in:

```
stage0    |     stage1      |     stage2 | ... | stageN      |     stageN+1 | ... | stageM
  X -> [copy] -> X_t0 -> [copy] -> X_t1 -> ... -> X_c0 -> [copy] -> X_tn+1 -> ... -> X_c1
```

Test Plan: Changes to `pipeline_discontiguous_ipu_copy_test_0.cpp` confirm that duplicate copies have been removed.

Reviewers: jamesn, simonl

Reviewed By: jamesn

Maniphest Tasks: T11256

Differential Revision: https://phabricator.sourcevertex.net/D15875