Added Atan, Asin, Sinh in popart
Summary:
I implemented the asin, atan, ans sinh operators in popart.
For the SINH and ATAN operators I used functions already implemented into poplar.
For ASIN I had to add the actual asin function to the POPLIBS list and then call it from here. I will add a different diff to poplibs for that and put the link into this task.
All the functions are implemented for outofplace, inplace and also the gradient

Test Plan: I added the tests for each of the operators into the operators_test.py script. I checked that they are all green.

Reviewers: jamesn, simonl, george

Reviewed By: jamesn

Maniphest Tasks: T6437

Differential Revision: https://phabricator.sourcevertex.net/D15721