Rename class to PopartTestSession
Summary:
The TestSession class in test_session.py was being misinterpreted as a test case when imported as its name began with "Test". This was throwing a warning on the buildbots and failing tests when run manually.

Just renaming the class fixes it. Fix incoming.

Test Plan: Tests pass

Reviewers: simonl, jamesn, evan, george

Reviewed By: simonl

Maniphest Tasks: T12404

Differential Revision: https://phabricator.sourcevertex.net/D16295