Add script that loads ir_dump.json into python.
Summary:
This makes it easier to use the json representation of the ir. Useful for debugging.

Useage:
  python3 -i scripts/load_ir_dump.py path_to_ir_dump.json

This will start an interactive python session with the ir dump loaded to a variable `ir`.
The main graph can be attained with `ir.main_graph()`.

To get the json representation of the ir, set the environment variable `POPART_IR_DUMP=ir_dump.json`. This will write the ir to `ir_dump.json` at the end of `Ir::prepare`, or when `Ir::prepare` throws an exception.

Reviewers: jamesn, evan, matthewha

Reviewed By: matthewha

Differential Revision: https://phabricator.sourcevertex.net/D16879