Another attempt to build PopART Python API doc
Summary:
In the last attempt, I miscalculated where the popart Python module was expecting to find the lib file. After checking this twice, I hope that I have got it right this time.

The python popart module is created in: ${PROJECT_BINARY_DIR}/python

The library file is also built in that directory

The python code (in __init__.py) in that directory attempts to load the library from: ../../lib (relative to itself)

Therefore, the library file should be copied to ${PROJECT_BINARY_DIR}/../lib (not ${PROJECT_BINARY_DIR}/lib as I had before)

Reviewers: #software_-_frameworks, george

Reviewed By: #software_-_frameworks, george

Maniphest Tasks: T11984

Differential Revision: https://phabricator.sourcevertex.net/D17143