Changed builder API so Ops that take a GraphProto argument now take a Builder reference instead
Summary:
This works around this issue that the popart cpp function to generate an `onnx::GraphProto` from a serialized model is not exposed in python.

It has the additional benefit that we no longer need to import onnx to use the python api when building Graphs with these Ops. Re-ran gen_opertators.py

Note: re-running `python gen_operators.py` has, in addition to changing the declarations and definitions of the `Loop`, `If`, and `Scan` ops, has re-shuffled the order of many other ops.
I have not been able to run this without this happening.

Test Plan: All tests pass (osx)

Reviewers: jamesn, george, matthewha, mohammeds

Reviewed By: mohammeds

Maniphest Tasks: T12857

Differential Revision: https://phabricator.sourcevertex.net/D17271