Add PatternNames class.
Summary:
Change the way pattern names are handled.
This adds the ability to get a pattern name without having an instance of the class via `PatternNames::getName<SomePattern>()`.
This also ensures all instances of patterns always have a name, where previously it was possible to construct a pattern and not call `initialise` to set the name.

Reviewers: jamesn, evan, simonl, matthewha

Reviewed By: jamesn

Maniphest Tasks: T11668

Differential Revision: https://phabricator.sourcevertex.net/D17476