Error message for non power of 2 IPUs.
Summary:
Tested and it's only possible to acquire a power of two IPUs. I didn't want to impose an upper limit in case we miss it when moving onto bigger machines.

Is this the only message we want? It's a bit confusing when `acquireAvailableDevice` returns nothing and you try and pass it to a session IMO.

Test Plan: Tests pass

Reviewers: jamesn, evan, george

Reviewed By: george

Tags: #popart

Maniphest Tasks: T12543

Differential Revision: https://phabricator.sourcevertex.net/D16691