Remove options.py and renamed SessionOptionsCore to SessionOptions.
Summary:
Throwing an error for a unrecognised attribute is the default behaviour with classes bound using pybind, as it is with python built-ins.
I have removed the python file, options.py, as it was just renaming `SessionOptionsCore` to `SessionOptions` using:
```
class SessionOptions : SessionOptionsCore:
    pass
```
with the side effect that this was allowing new attributes to be set on `SessionOptions`.

Test Plan: Added a test case to options_test.py

Reviewers: jamesn, evan, simonl, matthewha

Reviewed By: evan

Maniphest Tasks: T13127

Differential Revision: https://phabricator.sourcevertex.net/D17546