Fix gradient copy and variable copy scheduling for host side all reduce
Summary:
After trying to run host side all reduce with Horovod with a Resnet model I uncovered some bugs and a slightly more serious issue.

One of the bugs was that weight copies from the host were unordered with respect to when they were updated on the host.

A serious issue which was causing deadlocks was that Poplar ended up merging stream copies in which there was a dependency for the host read to complete before the host write was performed, even though the scheduling was correct in PopART.
I've left a code comment explaining this in `hostreducevarupdatex.cpp`

> Gradient copies are scheduled to happen before weight copies in PopART. However, if multiple stream copies are performed with a single sync id then a host read can be scheduled before a host write in the Poplar engine but the actual callback might still be executed after. This happens when Poplar merges two host syncs during compilation into one. See IPUTarget::prepareForStreamAccess() and IPUTarget::completeStreamAccess() for details

The remedy for this was to introduce a `poplar::program::Sync` after the gradient copy

Test Plan: All tests passing

Reviewers: jamesn, davidn, george

Reviewed By: jamesn

Differential Revision: https://phabricator.sourcevertex.net/D17869