Resolve T13558 (2 bug fixes, see below)
Summary:
RestoreInplace inserted for recomputation had no consumers, causing it to be
pruned / scheduled too late (I'm not sure). This diff makes the Op which is
recomputed using the Tensor it restores into a consumer. This does look a bit
odd in the Ir, but this is fundamentally because the Ir does not express
recompute.

DropoutGradOpx is called before DropoutOpx. I'm not entirely sure how
this is possible but I think it related to the changes introduced with the
SequenceMap logic where grow calls are postponed. This caused the reference
Tensor for the seed modifier in the backwards pass to be missing from the map.
Solution in this diff is to allow the DropoutGradOpx to create the reference
Tensor, in the same way as the forward Op can.

Test Plan: Test passes with topological tie-ing enabled in MatMulPattern. OTHER TESTS STILL RUNNING

Reviewers: evan, george, fabiant

Reviewed By: george

Differential Revision: https://phabricator.sourcevertex.net/D18723