Pytorch popart implementation attempt
Summary:
Just an idea of how to implement pytorch models in popart, done mostly for my understanding.

Wanted to use this example : https://pytorch.org/tutorials/beginner/blitz/cifar10_tutorial.html with as few changes as possible to the original code.
Now changes to feed the inputs and targets to the session. Details of why will follow in documentation.
The inputs and targets are now named better, and there is support for multiple inputs and targets, though it is assumed that there is one target for each output in the Nll case. Further improvements are probably necessary for identity loss support.

Added in loss and optimiser, for the couple of examples which we support.

Issues:
- User specifies "batch_size" as batch_size * batches_per_step in the dataloader. This is confusing.
- Anchor returns are the full N steps long. This is also confusing for a pytorch user.
- You have to specify batches_per_step in the session call.
- No support for multiple inputs / outputs / labels

Test Plan: The model trains!

Reviewers: evan, jamesn, george

Reviewed By: evan

Subscribers: scottg

Tags: #popart

Differential Revision: https://phabricator.sourcevertex.net/D17858