Fix the order of the patterns.
Summary:
The order of the patterns appears to be important, and changing the order is causing tests to fail on osx.
By sorting the patterns returned by `Patterns::getPreAliasList` by the corresponding `PreAliasPatternType`, the pattern order that existed before custom patterns is preserved.

This should not be viewed as a long term solution, but it returns popart to the state it was in pre custom patterns.
T14139 has been created to investigate the underlying issue.

Reviewers: evan, jamesn, matthewha

Reviewed By: evan

Differential Revision: https://phabricator.sourcevertex.net/D18826