Upgrading expected clang-format to 9.0.0, running format.sh on all code
Summary:
The default clang-format version (osx, ubuntu) installed with `brew install clang-format` is 9.0.0.

For new developers joining the team, it makes sense to use this version. I can't see a simple way to install older versions (8.0.0).

So I think we should update the version we use in line with this. You can make this change locally simply with `brew upgrade clang-format`

Reviewers: jamesn, george, matthewha

Reviewed By: matthewha

Subscribers: mohammeds

Differential Revision: https://phabricator.sourcevertex.net/D19348