Fix issue in SerializeMatMuls and add error check.
Summary:
There were instances where the `SerializeMatMul` transform was changing the output shape of some of the backward pass matmuls.
This diff adds a fix for this situation, and also adds a check in the transform that should error if this happens again.

Test Plan: As the running of other patterns was hiding the issues, the tests in `serializematmul.py` have been modified to only use the minimum set of patterns required for the test to run.

Reviewers: simonl, jamesn

Reviewed By: jamesn

Maniphest Tasks: T14638

Differential Revision: https://phabricator.sourcevertex.net/D19536