Sync Op implementation
Summary:
We need to enforce an internal sync in Poplar between the last gradient copy to host and the first gradient/weight copy to device due to stream copy merging in Poplar.

This was previously achieved with a stateful flag in devicex that monitored the `grow` operations. However since the `grow` functions are no longer necessarily called in the same sequence as the schedule we need to move the sync into a separate op and enforce the correct ordering with topocons.

Test Plan:
All tests pass.

New checks for correct op scheduling add to HostReduce unit tests

Reviewers: jamesn, george

Reviewed By: jamesn

Tags: #horovod, #popart

Maniphest Tasks: T14635

Differential Revision: https://phabricator.sourcevertex.net/D19572