Fix broken host reduce test.
Summary:
This fixes the failing osx buildbot.

Test previously tested model output against ground truth output computed by running the test with `opts.hostAllReduce = False`.
The test was initialized with a seeded random number generator, but this did not ensure the same random data across different platforms.

I have changed the test to run the model twice, with `hostAllReduce` set to true and false, and then compare the output of the two models.

Reviewers: jamesn, evan, matthewha

Reviewed By: jamesn

Differential Revision: https://phabricator.sourcevertex.net/D20069