Add option to serialise the ir without using the scheduler.
Summary: This option is used when dumping the ir after an error, as the graph may not always be schedulable at this point.

Test Plan: This is actually a really awkward one to write a test for. I have come across instances when an error has been thrown, and then when trying to dump the ir, another error is thrown by the scheduler, but I can't think of a way to force this behaviour.

Reviewers: jamesn, evan, matthewha

Reviewed By: matthewha

Maniphest Tasks: T14648

Differential Revision: https://phabricator.sourcevertex.net/D20146