Change dummy input name and add error check.
Summary:
Came across a situation where there was an op `Reshape(inputs=[clone_Reshape:0/3], outputs=[clone_Reshape:0/30])`.
The line, `t_clone->id += std::to_string(index_tensor.first);` evaluated to `"clone_Reshape:0/3" += "0";` so `clone_Reshape:0/30` which is the same name as the output.
First the input tensor was inserted into the map under this name, and then the output tensor was inserted with the same name, replacing the initial tensor, and causing the `unique_ptr<Tensor*>` instance to delete the input tensor.

I've added an extra '/' when adding the index to the tensor name, but also added error checks to prevent silently overwriting entries in `tensors`.

Reviewers: jamesn

Reviewed By: jamesn

Differential Revision: https://phabricator.sourcevertex.net/D20239