Add a ctest fixture for downloading the CIFAR10 data
Summary:
Ref T5216.

Previously the data was downloaded via a CTestCustom.ctest script and also each test had the ability to download it themselves. These two methods had a few downsides:
 - If `ctest` was not run in the same diretory as the custom script then it was not executed before the tests were run,
 - If the tests were run in parallel and they all tried to download the data to the same location it could get corrupted.

This change makes a test fixture that downloads the data and then all tests that use it require that fixture. It also removed their ability to download it themselves. As an added bonus this also means that tests that do not use the data will be run in parallel while it is being downloaded.

Test Plan: manually tests locally.

Reviewers: #popart, george

Reviewed By: #popart, george

Maniphest Tasks: T5216

Differential Revision: https://phabricator.sourcevertex.net/D20553