Use the same Python to run the tests that was used to build the library
Summary:
If not explicit the tests could, for example, try and run pytest via Python 2.7.

This would happen on the poplar buildbot as the build is not done inside of a virtualenv. Ref T5216.

Test Plan: can run tests locally without using a Python 3 virtualenv (was unable to before this change).

Reviewers: #popart, evan, jamesn, george, matthewha

Reviewed By: #popart, matthewha

Maniphest Tasks: T5216

Differential Revision: https://phabricator.sourcevertex.net/D20627