Support remote buffers for host side reductions
Summary:
We run into issues with HEXOPT size limitations when trying to run BERT with host side reductions.

By using `poplar::RemoteBuffer` we can get around the size limits. A drawback is that we need to introduce some hacks to force the IPU into a sync from which the host can read/write to the remotebuffers.

Test Plan:
New tests added which run when the machine supports OATT.

The unit test is getting bloated, I have added T15374 to refactor it which will most likely be done when a new hire is onboarded.

Reviewers: jamesn, george, #popart

Reviewed By: jamesn, #popart

Tags: #horovod, #popart

Maniphest Tasks: T15237

Differential Revision: https://phabricator.sourcevertex.net/D20624